Combine common package with dictionary
