+++ apps/server/package.json
@@ -33,13 +33,13 @@
 		"nodemon": "^2.0.22",
 		"prettier": "^3",
 		"testcontainers": "^1.3.1",
+		"typescript": "^5.5.4",
 		"zod-to-json-schema": "^3.21.3"
 	},
 	"dependencies": {
 		"ajv": "^8.12.0",
 		"axios": "^1.4.0",
 		"body-parser": "^1.20.2",
-		"common": "workspace:^",
 		"dictionary": "workspace:^",
 		"dotenv": "^16.3.1",
 		"errorhandler": "^1.5.1",
