+++ apps/server/src/controllers/dictionaryController.ts
@@ -17,8 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { BadRequestError, NotFoundError } from 'common';
-import { Dictionary, Schema, replaceReferences } from 'dictionary';
+import { BadRequestError, Dictionary, NotFoundError, Schema, replaceReferences } from 'dictionary';
 import { Request, Response } from 'express';
 import * as dictionaryService from '../services/dictionaryService';
