+++ packages/client/src/rest/index.ts
@@ -17,8 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { unknownToString } from 'common';
-import { Dictionary, DictionaryDiff, DictionaryDiffArray, FieldDiff } from 'dictionary';
+import { Dictionary, DictionaryDiff, DictionaryDiffArray, FieldDiff, unknownToString } from 'dictionary';
 import fetch from 'node-fetch';
 import promiseTools from 'promise-tools';
 import { loggerFor } from '../logger';
