--- packages/common/.mocharc.json
@@ -1,5 +1,0 @@
-{
-	"extension": ["ts"],
-	"require": "ts-node/register",
-	"spec": "test/**/*.spec.ts"
-}
