--- packages/common/README.md
@@ -1,5 +1,0 @@
-# Lectern Common Utils
-
-Contains generic types and utilities used across Lectern modules.
-
-Also exported are custom error classes. Please note: it is prefered to return Result objects with typed error details rather than throw unexpectedly. Use of thrown errors is being refactored out where possible.
