--- packages/common/package.json
@@ -1,16 +1,0 @@
-{
-	"name": "common",
-	"version": "1.0.0",
-	"description": "",
-	"main": "dist/index.js",
-	"scripts": {
-		"build": "pnpm build:clean && tsc",
-		"build:clean": "rimraf dist/ && mkdir dist"
-	},
-	"keywords": [],
-	"author": "Ontario Institute for Cancer Research",
-	"license": "AGPL-3.0",
-	"devDependencies": {
-		"rimraf": "^5.0.0"
-	}
-}
