--- packages/common/tsconfig.json
@@ -1,20 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ESNext",
-		"lib": ["ESNext"],
-		"module": "CommonJS",
-		"moduleResolution": "node",
-		"resolveJsonModule": true,
-		"strict": true,
-		"noImplicitAny": true,
-		"noUnusedParameters": true,
-		"esModuleInterop": true,
-		"declaration": true,
-		"sourceMap": true,
-		"inlineSources": true,
-		"outDir": "dist/",
-		"baseUrl": "./src",
-		"skipLibCheck": true
-	},
-	"include": ["./src/**/*.ts"]
-}
