+++ packages/dictionary/src/errors.ts
