+++ packages/dictionary/src/references.ts
@@ -19,8 +19,10 @@
 
 import * as immer from 'immer';
 import { cloneDeep, get, isObject, omit } from 'lodash';
-import { Dictionary, ReferenceArray, ReferenceTag, ReferenceValue, References, Schema } from '.';
-import { InvalidReferenceError, asArray } from 'common';
+import { Dictionary, ReferenceArray, ReferenceTag, ReferenceValue, References, Schema, TypeUtils } from '.';
+import { InvalidReferenceError } from './errors';
+
+const { asArray } = TypeUtils;
 
 // This is the union of all schema sections that could have reference values
 type OutputReferenceValues = ReferenceArray | ReferenceValue;
