+++ packages/dictionary/src/types/defined.ts
