+++ packages/dictionary/src/types/generics.ts
