+++ packages/dictionary/src/types/index.ts
@@ -17,10 +17,9 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './commonTypes';
-export * from './dataTypes';
-export * from './dbTypes';
-export * from './dictionaryTypes';
 export * from './diffTypes';
-export * from './referenceTypes';
-export * from './restrictionsTypes';
+
+export * from './defined';
+export * from './generics';
+export * from './result';
+export * from './singular';
