+++ packages/dictionary/src/types/result.ts
