+++ packages/dictionary/src/types/singular.ts
