+++ packages/dictionary/src/utils/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,23 +17,6 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { z as zod } from 'zod';
-import { DictionaryBase } from './dictionaryTypes';
-
-/**
- * A Dictionary stored in the DB is represented as a document and gets an `_id` property
- */
-export const DictionaryDocument = DictionaryBase.extend({
-	_id: zod.string(),
-	createdAt: zod.string(),
-	updatedAt: zod.string(),
-});
-export type DictionaryDocument = zod.infer<typeof DictionaryDocument>;
-
-export const DictionaryDocumentSummary = DictionaryDocument.pick({
-	name: true,
-	version: true,
-	description: true,
-	_id: true,
-});
-export type DictionaryDocumentSummary = zod.infer<typeof DictionaryDocumentSummary>;
+export * from './stringUtils';
+export * as TypeUtils from './typeUtils';
+export * as VersionUtils from './versionUtils';
