+++ packages/dictionary/src/utils/stringUtils.ts
