+++ packages/dictionary/src/utils/typeUtils.ts
