+++ packages/dictionary/src/utils/versionUtils.ts
@@ -17,7 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { MalformedVersionError } from 'common';
+import { MalformedVersionError } from '../errors';
 
 export const isValidVersion = (version: string): boolean => {
 	return /^[0-9]+\.[0-9]+$/.test(version);
