+++ packages/validation/src/validateSchema/validateSchema.ts
@@ -18,13 +18,13 @@
  */
 
 import type { DataRecord, Schema } from 'dictionary';
+import { TypeUtils } from 'dictionary';
 import { invalid, valid, type TestResult } from '../types';
-import type { SchemaValidationError, SchemaValidationRecordErrorDetails } from './SchemaValidationError';
 import { validateRecord } from '../validateRecord';
 import { generateDataSetHashMap } from './restrictions/generateDataSetHashMap';
-import { testUniqueKey } from './restrictions/uniqueKey/testUniqueKey';
-import { isDefined } from 'common';
 import { testUniqueFieldRestriction } from './restrictions/uniqueField/testUniqueFieldRestriction';
+import { testUniqueKey } from './restrictions/uniqueKey/testUniqueKey';
+import type { SchemaValidationError, SchemaValidationRecordErrorDetails } from './SchemaValidationError';
 
 /**
  * Validate a data set using a Lectern Schema. The data to validate is an array of DataRecords that contains all
@@ -82,6 +82,6 @@ export const validateSchema = (records: Array<DataRecord>, schema: Schema): Test
 			}
 			return recordErrors.length ? { recordIndex, recordErrors } : undefined;
 		})
-		.filter(isDefined);
+		.filter(TypeUtils.isDefined);
 	return schemaValidationErrors.length ? invalid(schemaValidationErrors) : valid();
 };
